pyproject.toml
[tool.poetry]
name = "pet-tracker"
version = "0.1.0"
description = ""
authors = ["Tom Whitwell <tom@whi.tw>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
Pillow = "^9.1.0"
arrow = "^1.2.2"
django-imagekit = "^4.1.0"
django-crispy-forms = "^1.14.0"
sweetify = { git = "https://github.com/joearton/sweetify-django.git", branch="patch-1"}
fontawesomefree = "^6.1.1"
django-adminlte3 = "^0.1.6"
django-polymorphic = "^3.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
autoflake = "^1.4"
ipython = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
